refactor: 포트 표현
