+++ worker/src/application/ports/sender.ts
@@ -1,4 +1,4 @@
-import { NotificationEntity } from "./dto";
+import { NotificationEntity } from "../dto";
 
 export interface INotificationSender {
     dispatch(notification: NotificationEntity[]): Promise<void>;
